+++ pkg/clusteragent/admission/mutate/autoinstrumentation/imageresolver/resolver_noop.go
@@ -1,0 +1,29 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build kubeapiserver
+
+package imageresolver
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/clusteragent/admission/metrics"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+)
+
+// noOpImageResolver is a simple implementation that returns the original image unchanged.
+// This is used when no remote config client is available.
+type noOpImageResolver struct{}
+
+// newNoOpImageResolver creates a new noOpImageResolver.
+func newNoOpImageResolver() ImageResolver {
+	return &noOpImageResolver{}
+}
+
+// ResolveImage returns the original image reference.
+func (r *noOpImageResolver) Resolve(registry string, repository string, tag string) (*ResolvedImage, bool) {
+	log.Debugf("Cannot resolve %s/%s:%s without remote config", registry, repository, tag)
+	metrics.ImageResolutionAttempts.Inc(registry, repository, metrics.DigestResolutionDisabled, tag)
+	return nil, false
+}
